% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{buildArchitecture}
\alias{buildArchitecture}
\title{Construct a VNNArchitecture from a GenePathwayMap}
\usage{
buildArchitecture(object, ...)
}
\arguments{
\item{object}{A \code{GenePathwayMap} object.}

\item{...}{Additional arguments passed to constructors.}
}
\value{
A \code{VNNArchitecture} object.
}
\description{
Construct a VNNArchitecture from a GenePathwayMap
}
