% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VNNBio-package.R
\docType{package}
\name{VNNBio-package}
\alias{VNNBio}
\alias{VNNBio-package}
\title{VNNBio: Visible Neural Networks for Biology}
\description{
Implements Visible Neural Networks (VNNs) whose architecture is
constrained by prior biological knowledge such as gene-pathway mappings.
Network connectivity mirrors known biology, so each hidden node corresponds
to an interpretable biological concept.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mxn581-dev/VNNBio}
  \item Report bugs at \url{https://github.com/mxn581-dev/VNNBio/issues}
}

}
\author{
\strong{Maintainer}: Minh Hieu Nguyen \email{mxn581@case.edu} (\href{https://orcid.org/0009-0009-6206-6121}{ORCID})

}
\keyword{internal}
